Improve secure context work
