+++ src/config/index.js
@@ -1,6 +1,5 @@
 import convict from 'convict'
 import path from 'path'
-import { getTrustStoreCerts } from '~/src/config/helpers/get-trust-store-certs'
 
 const oneWeek = 7 * 24 * 60 * 60 * 1000
 
@@ -59,11 +58,6 @@ const config = convict({
     format: ['fatal', 'error', 'warn', 'info', 'debug', 'trace', 'silent'],
     default: 'info',
     env: 'LOG_LEVEL'
-  },
-  trustStore: {
-    doc: 'CA Certificates',
-    format: Array,
-    default: getTrustStoreCerts(process.env)
   }
 })
