+++ src/server/common/helpers/secure-context/index.js
@@ -1,0 +1,3 @@
+import { secureContext } from '~/src/server/common/helpers/secure-context/secure-context'
+
+export { secureContext }
